initial work on resolving cyclic reference issue
