+++ core/shared/src/main/scala/zio/scheduler/package.scala
@@ -16,8 +16,41 @@
 
 package zio
 
+import java.util.concurrent.ScheduledExecutorService
+
+import zio.duration.Duration
+import zio.internal.IScheduler
+
 package object scheduler {
   type Scheduler = Has[Scheduler.Service]
+
+  object Scheduler extends PlatformSpecific {
+
+    trait Service extends Serializable {
+      def schedule[R, E, A](task: ZIO[R, E, A], duration: Duration): ZIO[R, E, A]
+    }
+
+    val defaultScheduler: Scheduler.Service =
+      fromIScheduler(globalScheduler)
+
+    val live: ZLayer.NoDeps[Nothing, Scheduler] =
+      ZLayer.succeed(defaultScheduler)
+
+    /**
+     * Creates a new `Scheduler` from a Java `ScheduledExecutorService`.
+     */
+    final def fromScheduledExecutorService(service: ScheduledExecutorService): Scheduler.Service =
+      fromIScheduler(IScheduler.fromScheduledExecutorService(service))
+
+    private[zio] def fromIScheduler(scheduler: IScheduler): Scheduler.Service =
+      new Scheduler.Service {
+        def schedule[R, E, A](task: ZIO[R, E, A], duration: Duration): ZIO[R, E, A] =
+          ZIO.effectAsyncInterrupt { cb =>
+            val canceler = scheduler.schedule(() => cb(task), duration)
+            Left(ZIO.effectTotal(canceler()))
+          }
+      }
+  }
 
   def scheduler: ZIO[Scheduler, Nothing, Scheduler.Service] =
     ZIO.access(_.get)
